I believe I've successfully worked around the ongoing performance problems associated with biz.saveAll() and biz.cancelAll() in the common case of calling these on a biz with no Parent but a tree of children below. The most common example is when the user clicks the 'save' button in a dForm.
The problem was that saveAll() and cancelAll() had to iterate every single
record, and every single record in all child bizobjs, to find changes. My
change only makes the entire iteration necessary in an exceptional case
if at all.

We now maintain a set() of keys of visited rows on the top-parent bizobj,
and during saveAll() and cancelAll() we first save/cancel changes to the
records we know we've visited since the last requery(), save(), or cancel(). 
Then, we call isAnyChanged() which will be lightning fast if there are no 
more changes, to determine if we need to do the old behavior of iterating all 
rows to catch the outliers.

Testing on my app yields huge performance gains (<1 second down from ~20) when 
modifying one field in one parent record, and clicking the save button.