Added biz.CascadeDeleteFromParent property, to allow different settings for different children.
Sometimes I make children for the purpose of having a convenient lookup, such
as:

	bizOrders.addChild(bizCustomers())

I wanted to keep this convenience but still be able to delete an order without
deleting the corresponding customer record. I couldn't just set the the 
deletechildlogic on bizOrders to "IGNORE" because there are children that I do
want to cascade the deletes to. 

So now I've changed that call to:

  bizOrders.addChild(bizCustomers(..., CascadeDeleteFromParent=False))