If the contents of a Memo overflows the available space (height is fixed too small), then an ellipse ("...") is printed after the length that can be shown to indicate truncated data on the report.
However, if that data had embedded html tags like <b>bold</b> or <i>italic</i>,
an error would be raised from reportlab if the string happened to get 
broken inside an open or closing tag (bare "<").

This commit fixes the problem by making sure the test string contains
matching tag brackets. 

Note: the <b> and <i> tags are acceptable in Memo fields to provide simple
in-string formatting for emphasis, so it wouldn't have worked to simply
replace the "<" with "&lt;".