Disabled appending "Hack: see hackDeferredPara()" to deferred memos, as it no longer seems necessary (because I'm seeing that string in my reports).
One of my reports just has basically one growable memo field in the
growable detail band. A blank page starts the report if the memo
won't all fit on the first page. Added logic for cases like this to
just start printing it anyway, knowing it'll flow to the next place
appropriately.

We don't want to do this in all cases because for memo fields with 
not very tall content, it would look stupid to print one line on 
this page and the second line on the next.