r7130 fixed a problem on Windows displaying the Euro symbol on Windows (UnicodeDecodeError coming up from reportlab). However, it turns out that the fix resulted in other unicode strings not displaying correctly (garbage characters added).
I'm not proud of this commit: I have no more time to try to work out what is happening,
but it does fix both problems. Basically, I reverted r7130 and then wrapped the 
reportlab draw() call in a try block catching for UnicodeDecodeError, and if it happens run
the string through dabo.lib.ustr().