Removed biz.Caption from the EditPage.Caption.
This is redundant since the form already has the caption, and looks better
because 1) it takes less horizontal space and 2) the biz.Caption is likely
plural, but in this context it should be singular. I didn't want to just
strip out a trailing "s" due to that being a locale-specific solution.