Fix _getID() to get it from wx.NewId().
Prior to this, we were filling dMenu id's using the logic from dObject,
which was to return the python id. This doesn't work in 64-bit Python
because wxPython (even 64-bit wxPython) doesn't recognize large ints
as ints, causing a StackOverflow exception.

This is better also since we really want wx to dole out the id's for
wxPython objects.