Settings changes for 1.0
This breaks backward compatibility by changing defaults for the
following settings:

  importDebugger: now defaults to False
  localizeDabo: now defaults to False
  implicitImports: now defaults to False

Each of these can easily be turned back to True in settings_override
if needed but also the needed changes aren't great. In the case of
localizing Dabo the code the appdev will need to add in their app
is:

  from dabo import dLocalize
  dLocalize.install("dabo")

Don't merge this into master until version 1.0.