Removed GridHeaderPaint event.
Raising it was keeping the grid headers from painting in wx2.9 on
Mac. I doubt anyone was relying on it but if so let me know before
we release Dabo 1.0.

I have a feeling most of the problems I've been fixing are more
related to Cocoa versus Carbon on Mac than to wxPython 2.9 versus
2.8.