Fixed grid cell editing/writing back to biz.
My last commit only handled bringing the value into the editor from
the bizobj, but not putting the value back correctly in the grid.
Specifically, it would change the type from date to string in the
bizobj when writing back.

Also discovered that we were redundantly saving back to the bizobj.
Fixed.