Fix conflict between splash and login dialog.
The splash screen and the login dialog both get shown before the
main event loop begins, and the login dialog is modal. Something
in this combination results in the login dialog becoming a child
of the splash screen, and therefore when the splash timeout is
released and the splash destroyed, it takes the login dialog with
it.

The solution is to defer destroying the splash screen until after
the login dialog has already been shown and hidden.

Another problem surfaced during this: when the login dialog was
hidden, the callAfter in dPemMixin._setVisible referred to
self.Parent.layout, but in this case self.Parent was a raw
wx.Frame and hence had no layout() method. Fixed by wrapping the
call in a try, catching AttributeError.

Tested on Linux with wxPython 2.8.12.1 and Python 2.7.3.