Ignore Name bindings in form in autoBindEvents.
My prior commit opened up allowing to bind on Name, but Ed correctly
pointed out that this would make no sense binding at the form level
to events from objects all over the hierarchy, potentially with the
same name.

With this change, the form won't bind to the event with a Name
reference except in the case that the object is a direct child of
the form.