Fixed artifacts when moving column with mouse.
On Linux at least, after moving a column with the mouse, the
header is placed correctly but the data is still lined up in the
old format (not aligned any longer with the header) until I
jiggle the grid. The immediate refresh call fixes it, but only
in concert with the delayed refresh call.