Sort-related minor enhancements.
Refactored code in dCursorMixin to use already-defined
function for moving to the old pk after the sort.

Added a form.update() after a grid sort so that the
row number is reflected correctly in the form's status
bar and potentially elsewhere.

Removed the check for rows in self.DataSet from the
gridHeaderMouseUp event handler, which greatly sped up
the sort processing time. Checking for DataSet when we
are using DataSource calls biz.getDataSet(). Pretty
misdirected of us to do that for no real reason.