Don't persist derived DataStructure.
The only time self._dataStructure should be saved is if it was
set explicitly by assigning to self.DataStructure. Otherwise,
it should be derived fresh in the first access after each requery.

I have a set of bizobjs where the user can add or remove fields
between requeries, and the DataStructure was staying the same
which wasn't expected.