Fix to layout() dLabel when the caption changes.
This has been bugging me for years and I finally did something
about it. When you change the caption of a dLabel, it isn't
enough to refresh() it, you need to call layout() on its Parent,
which I've been doing manually in my code and tending to forget
until I encounter it the next time.

Fixed to automatically call that parent.layout() when dLabel's
Caption changes.