Fix error with invisible columns in dGrid's autoSizeColumns.
When there are one or more invisible dColumns, the column count
in Dabo will be greater than the column count in wx (wx doesn't
know about the invisible columns). So if the user right-clicks
on the grid header and chooses to auto size all columns, wx
will raise an assertion error on Windows (doesn't do that on
Linux for some reason) that the column index is greater than the
column count.

This fixes it by first checking visibility, and then converting
the dabo column index to the wx column index.

There are probably other places in the codebase where this
conversion should be made as well (whenever sending a column
index down to wx).