Changed all tabs to spaces
Yes, I know that I was the one who pushed for the use of tabs, but that
was 15 years ago, when there was still some room for discussion.
Nowadays things have firmly settled on the side of spaces, and all my
tooling is set to automatically convert tabs to spaces, so I've thrown
in the towel, and henceforth Dabo will use spaces.

I also remove the DOS-style CRLFs.