Remove all the namespace mangling in the UI
Because of the previous design to accommodate multiple UI
implementations, we imported every control class into the dabo.ui
namespace. Removing the toolkit subdirectories means that now the
control modules are in the dabo/ui directory. This makes import
confusing, since the class names are the same as the module names: does
something like 'dabo.ui.dTextBox' refer to the dTextBox module in the
dabo/ui directory? Or does it refer to the dTextBox class defined in
that module?

This patch removes the class names from the dabo.ui namespace, and tries
to fix as many of the imports and references as possible. There are way
too many for one commit, so my goal is to be able to run the dEditor
class. After that a subsequent patch will clean up more.