Rework __init__() calls to use super()
wxPython changed the way that their objects check for "readiness", and
our old method of selectively calling the __init__()'s of superclasses
was causing wx to fail. So I've started changing them all to using
super(), and making sure that those method signatures match.

There is still a lot of debug printout in the code, but this is as far
as I got over the weekend.