Changed all the __init__() superclass calls
Dabo was created with very specific calls to superclasses, since the
Dabo-level controls required different parameters than the wx-level
classes. Apparently that won't work with wxPython 4.x, which requires
that super() be used or it will error out. So this is a gross change
replacing all those individual class-based calls to __init__() with the
super() version. There might be some conflicts, as this hasn't been
fully tested due to other issues, but it's a start.