A bunch of cleanups
I haven't had time to focus on working on this methodically, so instead
I've been running various things and fixing them when they are broken.

There are major problems with drawn objects - it seems that the way that
DCs work has changed quite a bit. Next time I get a chance I will look
into that.