I was getting lot's of these messages: DeprecationWarning: an integer is required (got type float). Implicit conversion to integers using __int__ is deprecated, and may be removed in a future version of Python.
I replaced the floats with a convert to int() to fix the above issues.  I'm not sure this the
best thing to do but I can't see any issues.

The ReportDesigner opens files, saves, the property sheet opens,
the object tree opens. The objects can be moved using the arrow keys.
Dragging does NOT work.  The object can't select an object.

ReportDesigner also works with the latest reportlab 3.5.49