Moved the demo code to a separate repo, and lots of other cleanup.
Most things are working, but not all very well. Grids work, but their
headers are not showing up. Mouse events are not all there: clicks work,
but scroll events are not working, at least on my Mac.

I will be moving the `ide` stuff into a separate repo soon. I thought I
could integrate it into the framework, but it is an application that
uses Dabo, not a part of Dabo itself.

Got the object inspector working by capturing the code created from the
.cdxml file and tweaking it. I'm not sure that the class designer stuff
is workable for the time being.