Updated all property definitions
The standard in the original version of Dabo was to define properties by
first creating getter and setter methods, and then calling the built-in
`property()` function on them. Example;

    def _getSomething(self):
        return self._something

    def _setSomething(self, val):
        self._something = val

    Something = property(self._getSomething, self._setSomething, None,
        "Docstring for this property")

That was the standard for that time. Now, it is more Pythonic to use the
decorator syntax, so all of these old-style definitions have been
updated to use the decorator style. The example above will now look
like:

    @property
    def Something(self):
        """Docstring for this property"""
        return self._something

    @Something.setter
    def Something(self, val):
        self._something = val

This process revealed some poor designs, where classes called the
super() of the getter/setter. These were updated to reference the
property if possible. For cases where that was not possible, new
functions were added to `dabo.lib.utils` to get the superclass
properties or set their value.