Change encryption to use the cryptography library
Previously Dabo relied on a simple cipher routine called SimpleCrypt
that was not secure at all. If you installed the PyCrypto library, Dabo
would detect that and use its encryption instead.

Since those early days, PyCrypto has been put in maintenance mode and
has been found to have several vulnerabilities. So it was time to remove
it from Dabo.

The `cryptography` library is a modern tool for handling cryptographic
needs. It has been added as a dependency, and both the SimpleCrypt and
PyCrypto code has been removed. Your application will need to supply a
password and an optional salt value. It will work without a salt, but
it will be less secure. There is also a default mode with no password or
salt that is needed for things like the Preference Manager that uses
local SQLite dbs that are not password-protected, but using this for any
other secrets is not recommended.

It is very important to not store neither your password nor salt in your
code. Instead, create a function in your code that will retrieve it from
a local file that is *NOT* included in your project. This way you can
use public sites like Github or Gitlab without anyone being able to
discover your secret values.