Fixed encryption when running outside an Application
Attempting to use a .cnxml file from an interpreter would crash with an
invalid token message. It was because we always set a key in the
dConnectInfo class, whether one was passed in or not, and setting the
key to None resulted in a default value being assigned, which did not
match the values used to encrypt the password originally.